fix(inventory): цена в БД — integer, без копеек
decimal(10,2) расходился с интерфейсом (step="1", number_format(..., 0)):
сервер принимал 2500.50, список округлял до «2 501 ₸», а модалка подставляла
2500.5 в поле со step="1" — браузер отказывался отправлять форму, и позицию
нельзя было ни переименовать, ни выключить.

Колонка integer + каст integer — как у ClubCardType::price. Таблица пустая,
миграция ещё не применена ни локально, ни на проде, так что правим на месте.

+++ database/migrations/2026_08_08_000001_create_club_inventory_items_table.php
@@ -12,7 +12,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('club_id')->constrained()->cascadeOnDelete();
             $table->string('name');
-            $table->decimal('price', 10, 2);
+            // Цена в целых тенге, без копеек — как в остальных денежных полях проекта.
+            $table->integer('price');
             $table->boolean('is_active')->default(true);
             $table->timestamps();
             // По этой паре строится список раздела.
